Remove v1 theme
