+++ packages/common/src/services/remote-config/feature-flags.ts
@@ -21,7 +21,6 @@ export enum FeatureFlags {
   RIGHTS_AND_COVERS = 'rights_and_covers',
   GUEST_CHECKOUT = 'guest_checkout',
   TRACK_AUDIO_REPLACE = 'track_audio_replace',
-  THEME_V2 = 'theme_v2',
   OWN_YOUR_FANS = 'own_your_fans',
   FAST_REFERRAL = 'fast_referral'
 }
@@ -60,7 +59,6 @@ export const flagDefaults: FlagDefaults = {
   [FeatureFlags.SOCIAL_SIGNUP]: false,
   [FeatureFlags.RIGHTS_AND_COVERS]: false,
   [FeatureFlags.TRACK_AUDIO_REPLACE]: false,
-  [FeatureFlags.THEME_V2]: false,
   [FeatureFlags.OWN_YOUR_FANS]: false,
   [FeatureFlags.FAST_REFERRAL]: false
 }
