--- packages/harmony/src/foundations/color/primitive-v2.css
@@ -1,222 +1,0 @@
-/**
- * Audius Color Palette V2
- */
-
-/* Theme Color Definitions */
-:root {
-  /* Static Colors */
-  --harmony-static-static-white: #ffffffff;
-  --harmony-static-black: #000000;
-  --harmony-static-primary: #cc0fe0ff;
-
-  /* Default Theme (Day Theme) */
-  /* Primary Colors */
-  --harmony-primary: #cc0fe0ff;
-  --harmony-p-100: #d63fe6ff;
-  --harmony-p-200: #d127e3ff;
-  --harmony-p-300: #cc0fe0ff;
-  --harmony-p-400: #b80ecaff;
-  --harmony-p-500: #a30cb3ff;
-
-  /* Secondary Colors */
-  --harmony-secondary: #7e1bccff;
-  --harmony-s-100: #9849d6ff;
-  --harmony-s-200: #8b32d1ff;
-  --harmony-s-300: #7e1bccff;
-  --harmony-s-400: #7118b8ff;
-  --harmony-s-500: #6516a3ff;
-
-  /* Neutral Colors */
-  --harmony-neutral: #6c6780ff;
-  --harmony-n-25: #fafafaff;
-  --harmony-n-50: #f7f7f8ff;
-  --harmony-n-100: #efeff1ff;
-  --harmony-n-150: #e7e7eaff;
-  --harmony-n-200: #d6d5dbff;
-  --harmony-n-300: #c3c1cbff;
-  --harmony-n-400: #a2a0afff;
-  --harmony-n-500: #938fa3ff;
-  --harmony-n-600: #817d95ff;
-  --harmony-n-700: #736e88ff;
-  --harmony-n-800: #6c6780ff;
-  --harmony-n-900: #615d73ff;
-  --harmony-n-950: #524f62ff;
-
-  /* Special Colors */
-  --harmony-white: #ffffffff;
-  --harmony-background: #f6f5f7ff;
-  --harmony-blue: #1ba1f1ff;
-  --harmony-orange: #ff9400ff;
-  --harmony-red: #d0021bff;
-  --harmony-dark-red: #bb0218ff;
-  --harmony-green: #0f9e48ff;
-  --harmony-light-green: #13c65aff;
-  --harmony-gradient-color-1: #5b23e1ff;
-  --harmony-gradient-color-2: #a22febff;
-  --harmony-gradient: linear-gradient(315deg, #5b23e1ff 0%, #a22febff 100%);
-  --harmony-trending-blue: #216fdcff;
-  --harmony-ai-green: #1fd187ff;
-  --harmony-glass-overlay: #ffffffd9;
-}
-
-/* Day Theme */
-html[data-theme='day-v2'] {
-  /* Primary Colors */
-  --harmony-primary: #cc0fe0ff;
-  --harmony-p-100: #d63fe6ff;
-  --harmony-p-200: #d127e3ff;
-  --harmony-p-300: #cc0fe0ff;
-  --harmony-p-400: #b80ecaff;
-  --harmony-p-500: #a30cb3ff;
-
-  /* Secondary Colors */
-  --harmony-secondary: #7e1bccff;
-  --harmony-s-100: #9849d6ff;
-  --harmony-s-200: #8b32d1ff;
-  --harmony-s-300: #7e1bccff;
-  --harmony-s-400: #7118b8ff;
-  --harmony-s-500: #6516a3ff;
-
-  /* Neutral Colors */
-  --harmony-neutral: #6c6780ff;
-  --harmony-n-25: #fafafaff;
-  --harmony-n-50: #f7f7f8ff;
-  --harmony-n-100: #efeff1ff;
-  --harmony-n-150: #e7e7eaff;
-  --harmony-n-200: #d6d5dbff;
-  --harmony-n-300: #c3c1cbff;
-  --harmony-n-400: #a2a0afff;
-  --harmony-n-500: #938fa3ff;
-  --harmony-n-600: #817d95ff;
-  --harmony-n-700: #736e88ff;
-  --harmony-n-800: #6c6780ff;
-  --harmony-n-900: #615d73ff;
-  --harmony-n-950: #524f62ff;
-
-  /* Special Colors */
-  --harmony-white: #ffffffff;
-  --harmony-background: #f6f5f7ff;
-  --harmony-blue: #1ba1f1ff;
-  --harmony-orange: #ff9400ff;
-  --harmony-red: #d0021bff;
-  --harmony-dark-red: #bb0218ff;
-  --harmony-green: #0f9e48ff;
-  --harmony-light-green: #13c65aff;
-  --harmony-gradient-color-1: #5b23e1ff;
-  --harmony-gradient-color-2: #a22febff;
-  --harmony-gradient: linear-gradient(315deg, #5b23e1ff 0%, #a22febff 100%);
-  --harmony-trending-blue: #216fdcff;
-  --harmony-ai-green: #1fd187ff;
-  --harmony-glass-overlay: #ffffffd9;
-
-  /* Overrides static-white to be regular white */
-  --harmony-static-white: var(--harmony-white);
-}
-
-/* Dark Theme */
-html[data-theme='dark-v2'] {
-  /* Primary Colors */
-  --harmony-primary: #d767e1ff;
-  --harmony-p-100: #924699ff;
-  --harmony-p-200: #b356bbff;
-  --harmony-p-300: #d767e1ff;
-  --harmony-p-400: #f479ffff;
-  --harmony-p-500: #f7a4ffff;
-
-  /* Secondary Colors */
-  --harmony-secondary: #c67cffff;
-  --harmony-s-100: #9a66caff;
-  --harmony-s-200: #b071e4ff;
-  --harmony-s-300: #c67cffff;
-  --harmony-s-400: #cf90ffff;
-  --harmony-s-500: #d7a3ffff;
-
-  /* Neutral Colors */
-  --harmony-neutral: #c6c8d6ff;
-  --harmony-n-25: #30354aff;
-  --harmony-n-50: #353a51ff;
-  --harmony-n-100: #3c425dff;
-  --harmony-n-150: #444b69ff;
-  --harmony-n-200: #4f577aff;
-  --harmony-n-300: #656f9bff;
-  --harmony-n-400: #7b83a7ff;
-  --harmony-n-500: #969bb6ff;
-  --harmony-n-600: #a6aac0ff;
-  --harmony-n-700: #b6b9cbff;
-  --harmony-n-800: #c6c8d6ff;
-  --harmony-n-900: #d6d7e1ff;
-  --harmony-n-950: #e6e7edff;
-
-  /* Special Colors */
-  --harmony-white: #2f3348ff;
-  --harmony-background: #202131ff;
-  --harmony-blue: #58b9f4ff;
-  --harmony-orange: #efa947ff;
-  --harmony-red: #f9344cff;
-  --harmony-dark-red: #c43047ff;
-  --harmony-green: #6cdf44ff;
-  --harmony-light-green: #15d864ff;
-  --harmony-gradient-color-1: #9469eeff;
-  --harmony-gradient-color-2: #c781fcff;
-  --harmony-gradient: linear-gradient(315deg, #9469eeff 0%, #c781fcff 100%);
-  --harmony-trending-blue: #2a85e8ff;
-  --harmony-ai-green: #1ecf7aff;
-  --harmony-glass-overlay: #2f334899;
-
-  /* Overrides static-white to be regular white */
-  --harmony-static-white: var(--harmony-white);
-}
-
-/* Matrix Theme */
-html[data-theme='matrix-v2'] {
-  /* Primary Colors */
-  --harmony-primary: #0bd90bff;
-  --harmony-p-100: #08a008ff;
-  --harmony-p-200: #09b509ff;
-  --harmony-p-300: #0bd90bff;
-  --harmony-p-400: #0be40bff;
-  --harmony-p-500: #3df23dff;
-
-  /* Secondary Colors */
-  --harmony-secondary: #00e252ff;
-  --harmony-s-100: #00a73cff;
-  --harmony-s-200: #00bd44ff;
-  --harmony-s-300: #00e252ff;
-  --harmony-s-400: #00ed56ff;
-  --harmony-s-500: #3cfb6bff;
-
-  /* Neutral Colors */
-  --harmony-neutral: #3ef56bff;
-  --harmony-n-25: #010b06ff;
-  --harmony-n-50: #041308ff;
-  --harmony-n-100: #0a1f0dff;
-  --harmony-n-150: #0f2f15ff;
-  --harmony-n-200: #123f1bff;
-  --harmony-n-300: #166527ff;
-  --harmony-n-400: #1c8736ff;
-  --harmony-n-500: #1f9e3fff;
-  --harmony-n-600: #2eb954ff;
-  --harmony-n-700: #34d561ff;
-  --harmony-n-800: #3ef56bff;
-  --harmony-n-900: #54ff80ff;
-  --harmony-n-950: #9bffa7ff;
-
-  /* Special Colors */
-  --harmony-white: #020804ff;
-  --harmony-background: #000000ff;
-  --harmony-blue: #20e290ff;
-  --harmony-orange: #ffa524ff;
-  --harmony-red: #f9344cff;
-  --harmony-dark-red: #d63c5aff;
-  --harmony-green: #2eb954ff;
-  --harmony-light-green: #0bf90bff;
-  --harmony-gradient-color-1: #6cdf44ff;
-  --harmony-gradient-color-2: #13c65aff;
-  --harmony-gradient: linear-gradient(315deg, #6cdf44ff 0%, #13c65aff 100%);
-  --harmony-trending-blue: #1eeb6fff;
-  --harmony-ai-green: #1eeb57ff;
-  --harmony-glass-overlay: #01010199;
-
-  /* Overrides static-white to be regular white */
-  --harmony-static-white: var(--harmony-white);
-}
