+++ packages/harmony/src/foundations/color/primitive.ts
@@ -1,162 +1,188 @@
 export const primitiveTheme = {
   day: {
     static: {
-      white: '#FFFFFF',
-      staticWhite: '#FFFFFF',
-      black: '#000000',
-      primary: '#CC0FE0'
+      primary: '#CC0FE0FF',
+      secondary: '#7E1BCCFF',
+      white: '#FFFFFFFF',
+      staticWhite: '#FFFFFFFF',
+      black: '#000000'
     },
     primary: {
-      primary: '#CC0FE0',
-      p100: '#D63FE6',
-      p200: '#D127E3',
-      p300: '#CC0FE0',
-      p400: '#B80ECA',
-      p500: '#A30CB3'
+      p100: '#D63FE6FF',
+      p200: '#D127E3FF',
+      p300: '#CC0FE0FF',
+      p400: '#B80ECAFF',
+      p500: '#A30CB3FF',
+      default: '#CC0FE0FF',
+      primary: '#CC0FE0FF'
     },
     secondary: {
-      secondary: '#7E1BCC',
-      s100: '#9849D6',
-      s200: '#8B32D1',
-      s300: '#7E1BCC',
-      s400: '#7118B8',
-      s500: '#6516A3'
+      s100: '#9849D6FF',
+      s200: '#8B32D1FF',
+      s300: '#7E1BCCFF',
+      s400: '#7118B8FF',
+      s500: '#6516A3FF',
+      default: '#7E1BCCFF',
+      secondary: '#7E1BCCFF'
     },
     neutral: {
-      neutral: '#858199',
-      n25: '#FCFCFC',
-      n50: '#F7F7F9',
-      n100: '#F2F2F4',
-      n150: '#E7E6EB',
-      n200: '#DAD9E0',
-      n300: '#CECDD6',
-      n400: '#C2C0CC',
-      n500: '#B6B3C2',
-      n600: '#AAA7B8',
-      n700: '#9D9AAD',
-      n800: '#858199',
-      n900: '#78748A',
-      n950: '#6A677A'
+      n25: '#FAFAFAFF',
+      n50: '#F7F7F8FF',
+      n100: '#EFEFF1FF',
+      n150: '#E7E7EAFF',
+      n200: '#D6D5DBFF',
+      n300: '#C3C1CBFF',
+      n400: '#A2A0AFFF',
+      n500: '#938FA3FF',
+      n600: '#817D95FF',
+      n700: '#736E88FF',
+      n800: '#6C6780FF',
+      n900: '#615D73FF',
+      n950: '#524F62FF',
+      default: '#6C6780FF',
+      neutral: '#6C6780FF'
     },
     special: {
-      white: '#FFFFFF',
-      background: '#F3F0F7',
-      blue: '#1BA1F1',
-      orange: '#FF9400',
-      red: '#D0021B',
-      darkRed: '#BB0218',
-      green: '#0F9E48',
-      lightGreen: '#13C65A',
-      trendingBlue: '#216FDC',
-      aiGreen: '#1FD187',
-      gradient: 'linear-gradient(315deg, #5b23e1 0%, #a22feb 100%)'
+      aiGreen: '#1FD187FF',
+      background: '#F6F5F7FF',
+      blue: '#1BA1F1FF',
+      darkRed: '#BB0218FF',
+      glassOverlay: '#FFFFFFD9',
+      gradientStop1: '#5B23E1FF',
+      gradientStop2: '#A22FEBFF',
+      gradient: 'linear-gradient(315deg, #5B23E1FF 0%, #A22FEBFF 100%)',
+      green: '#0F9E48FF',
+      lightGreen: '#13C65AFF',
+      orange: '#FF9400FF',
+      red: '#D0021BFF',
+      trendingBlue: '#216FDCFF',
+      white: '#FFFFFFFF',
+      default: '#1BA1F1FF',
+      special: '#1BA1F1FF'
     }
   },
-
   dark: {
     static: {
-      white: '#FFFFFF',
-      staticWhite: '#FFFFFF',
-      black: '#000000',
-      primary: '#CC0FE0'
+      primary: '#B94CC2FF',
+      secondary: '#7E1BCCFF',
+      white: '#2F3348FF',
+      staticWhite: '#F6F8FCFF',
+      black: '#000000'
     },
     primary: {
-      primary: '#C74BD3',
-      p100: '#A945B9',
-      p200: '#B748C6',
-      p300: '#C74BD3',
-      p400: '#C556D4',
-      p500: '#C563D6'
+      p100: '#924699FF',
+      p200: '#B356BBFF',
+      p300: '#D767E1FF',
+      p400: '#F479FFFF',
+      p500: '#F7A4FFFF',
+      default: '#D767E1FF',
+      primary: '#D767E1FF'
     },
     secondary: {
-      secondary: '#9147CC',
-      s100: '#7440A4',
-      s200: '#8244B8',
-      s300: '#9147CC',
-      s400: '#975ACD',
-      s500: '#9A60CF'
+      s100: '#9A66CAFF',
+      s200: '#B071E4FF',
+      s300: '#C67CFFFF',
+      s400: '#CF90FFFF',
+      s500: '#D7A3FFFF',
+      default: '#C67CFFFF',
+      secondary: '#C67CFFFF'
     },
     neutral: {
-      neutral: '#BEC5E0',
-      n25: '#35364F',
-      n50: '#393A54',
-      n100: '#3F415B',
-      n150: '#4E4F6A',
-      n200: '#5A5E78',
-      n300: '#696D88',
-      n400: '#777C96',
-      n500: '#868AA4',
-      n600: '#9399B3',
-      n700: '#A2A8C2',
-      n800: '#BEC5E0',
-      n900: '#CFD5EC',
-      n950: '#D9DEF1'
+      n25: '#30354AFF',
+      n50: '#353A51FF',
+      n100: '#3C425DFF',
+      n150: '#444B69FF',
+      n200: '#4F577AFF',
+      n300: '#656F9BFF',
+      n400: '#7B83A7FF',
+      n500: '#969BB6FF',
+      n600: '#A6AAC0FF',
+      n700: '#B6B9CBFF',
+      n800: '#C6C8D6FF',
+      n900: '#D6D7E1FF',
+      n950: '#E6E7EDFF',
+      default: '#C6C8D6FF',
+      neutral: '#C6C8D6FF'
     },
     special: {
-      white: '#32334D',
-      background: '#242438',
-      blue: '#58B9F4',
-      orange: '#EFA947',
-      red: '#F9344C',
-      darkRed: '#C43047',
-      green: '#6CDF44',
-      lightGreen: '#13C65A',
-      trendingBlue: '#216FDC',
-      aiGreen: '#1FD187',
-      gradient: 'linear-gradient(315deg, #7652cc 0%, #b05ce6 100%)'
+      aiGreen: '#1ECF7AFF',
+      background: '#202131FF',
+      blue: '#58B9F4FF',
+      darkRed: '#C43047FF',
+      glassOverlay: '#2F334899',
+      gradientStop1: '#9469EEFF',
+      gradientStop2: '#C781FCFF',
+      gradient: 'linear-gradient(315deg, #9469EEFF 0%, #C781FCFF 100%)',
+      green: '#6CDF44FF',
+      lightGreen: '#15D864FF',
+      orange: '#EFA947FF',
+      red: '#F9344CFF',
+      trendingBlue: '#2A85E8FF',
+      white: '#2F3348FF',
+      default: '#58B9F4FF',
+      special: '#58B9F4FF'
     }
   },
   matrix: {
     static: {
-      white: '#FFFFFF',
-      staticWhite: '#FFFFFF',
-      black: '#000000',
-      primary: '#CC0FE0'
+      primary: '#B94CC2FF',
+      secondary: '#7E1BCCFF',
+      white: '#020804FF',
+      staticWhite: '#F6F8FCFF',
+      black: '#000000'
     },
     primary: {
-      primary: '#0CF10C',
-      p100: '#0CF10C',
-      p200: '#0CF10C',
-      p300: '#0CF10C',
-      p400: '#0CF10C',
-      p500: '#0CF10C'
+      p100: '#08A008FF',
+      p200: '#09B509FF',
+      p300: '#0BD90BFF',
+      p400: '#0BE40BFF',
+      p500: '#3DF23DFF',
+      default: '#0BD90BFF',
+      primary: '#0BD90BFF'
     },
     secondary: {
-      secondary: '#184F17',
-      s100: '#184F17',
-      s200: '#184F17',
-      s300: '#184F17',
-      s400: '#184F17',
-      s500: '#184F17'
+      s100: '#00A73CFF',
+      s200: '#00BD44FF',
+      s300: '#00E252FF',
+      s400: '#00ED56FF',
+      s500: '#3CFB6BFF',
+      default: '#00E252FF',
+      secondary: '#00E252FF'
     },
     neutral: {
-      neutral: '#21B404',
-      n25: '#1D211B',
-      n50: '#202A1D',
-      n100: '#1A2F15',
-      n150: '#1B3714',
-      n200: '#1C5610',
-      n300: '#1D5E0F',
-      n400: '#1D660E',
-      n500: '#1F850A',
-      n600: '#1F9508',
-      n700: '#20A406',
-      n800: '#21B404',
-      n900: '#21B404',
-      n950: '#21B404'
+      n25: '#010B06FF',
+      n50: '#041308FF',
+      n100: '#0A1F0DFF',
+      n150: '#0F2F15FF',
+      n200: '#123F1BFF',
+      n300: '#166527FF',
+      n400: '#1C8736FF',
+      n500: '#1F9E3FFF',
+      n600: '#2EB954FF',
+      n700: '#34D561FF',
+      n800: '#3EF56BFF',
+      n900: '#54FF80FF',
+      n950: '#9BFFA7FF',
+      default: '#3EF56BFF',
+      neutral: '#3EF56BFF'
     },
     special: {
-      white: '#1F211F',
-      background: '#191818',
-      blue: '#58B9F4',
-      orange: '#EFA947',
-      red: '#F9344C',
-      darkRed: '#C43047',
-      green: '#6CDF44',
-      lightGreen: '#13C65A',
-      trendingBlue: '#216FDC',
-      aiGreen: '#1FD187',
-      gradient: 'linear-gradient(323.08deg, #4FF069 36.13%, #09BD51 133.51%)'
+      aiGreen: '#1EEB57FF',
+      background: '#000000FF',
+      blue: '#20E290FF',
+      darkRed: '#D63C5AFF',
+      glassOverlay: '#01010199',
+      gradientStop1: '#6CDF44FF',
+      gradientStop2: '#13C65AFF',
+      gradient: 'linear-gradient(315deg, #6CDF44FF 0%, #13C65AFF 100%)',
+      green: '#2EB954FF',
+      lightGreen: '#0BF90BFF',
+      orange: '#FFA524FF',
+      red: '#F9344CFF',
+      trendingBlue: '#1EEB6FFF',
+      white: '#020804FF',
+      default: '#20E290FF',
+      special: '#20E290FF'
     }
   }
 }
@@ -166,3 +192,4 @@ export type PrimaryColors = keyof PrimitiveColors['primary']
 export type SecondaryColors = keyof PrimitiveColors['secondary']
 export type NeutralColors = keyof PrimitiveColors['neutral']
 export type SpecialColors = keyof PrimitiveColors['special']
+export type StaticColors = keyof PrimitiveColors['static']
