--- packages/harmony/src/foundations/color/primitiveV2.ts
@@ -1,195 +1,0 @@
-export const primitiveThemeV2 = {
-  day: {
-    static: {
-      primary: '#CC0FE0FF',
-      secondary: '#7E1BCCFF',
-      white: '#FFFFFFFF',
-      staticWhite: '#FFFFFFFF',
-      black: '#000000'
-    },
-    primary: {
-      p100: '#D63FE6FF',
-      p200: '#D127E3FF',
-      p300: '#CC0FE0FF',
-      p400: '#B80ECAFF',
-      p500: '#A30CB3FF',
-      default: '#CC0FE0FF',
-      primary: '#CC0FE0FF'
-    },
-    secondary: {
-      s100: '#9849D6FF',
-      s200: '#8B32D1FF',
-      s300: '#7E1BCCFF',
-      s400: '#7118B8FF',
-      s500: '#6516A3FF',
-      default: '#7E1BCCFF',
-      secondary: '#7E1BCCFF'
-    },
-    neutral: {
-      n25: '#FAFAFAFF',
-      n50: '#F7F7F8FF',
-      n100: '#EFEFF1FF',
-      n150: '#E7E7EAFF',
-      n200: '#D6D5DBFF',
-      n300: '#C3C1CBFF',
-      n400: '#A2A0AFFF',
-      n500: '#938FA3FF',
-      n600: '#817D95FF',
-      n700: '#736E88FF',
-      n800: '#6C6780FF',
-      n900: '#615D73FF',
-      n950: '#524F62FF',
-      default: '#6C6780FF',
-      neutral: '#6C6780FF'
-    },
-    special: {
-      aiGreen: '#1FD187FF',
-      background: '#F6F5F7FF',
-      blue: '#1BA1F1FF',
-      darkRed: '#BB0218FF',
-      glassOverlay: '#FFFFFFD9',
-      gradientStop1: '#5B23E1FF',
-      gradientStop2: '#A22FEBFF',
-      gradient: 'linear-gradient(315deg, #5B23E1FF 0%, #A22FEBFF 100%)',
-      green: '#0F9E48FF',
-      lightGreen: '#13C65AFF',
-      orange: '#FF9400FF',
-      red: '#D0021BFF',
-      trendingBlue: '#216FDCFF',
-      white: '#FFFFFFFF',
-      default: '#1BA1F1FF',
-      special: '#1BA1F1FF'
-    }
-  },
-  dark: {
-    static: {
-      primary: '#B94CC2FF',
-      secondary: '#7E1BCCFF',
-      white: '#2F3348FF',
-      staticWhite: '#F6F8FCFF',
-      black: '#000000'
-    },
-    primary: {
-      p100: '#924699FF',
-      p200: '#B356BBFF',
-      p300: '#D767E1FF',
-      p400: '#F479FFFF',
-      p500: '#F7A4FFFF',
-      default: '#D767E1FF',
-      primary: '#D767E1FF'
-    },
-    secondary: {
-      s100: '#9A66CAFF',
-      s200: '#B071E4FF',
-      s300: '#C67CFFFF',
-      s400: '#CF90FFFF',
-      s500: '#D7A3FFFF',
-      default: '#C67CFFFF',
-      secondary: '#C67CFFFF'
-    },
-    neutral: {
-      n25: '#30354AFF',
-      n50: '#353A51FF',
-      n100: '#3C425DFF',
-      n150: '#444B69FF',
-      n200: '#4F577AFF',
-      n300: '#656F9BFF',
-      n400: '#7B83A7FF',
-      n500: '#969BB6FF',
-      n600: '#A6AAC0FF',
-      n700: '#B6B9CBFF',
-      n800: '#C6C8D6FF',
-      n900: '#D6D7E1FF',
-      n950: '#E6E7EDFF',
-      default: '#C6C8D6FF',
-      neutral: '#C6C8D6FF'
-    },
-    special: {
-      aiGreen: '#1ECF7AFF',
-      background: '#202131FF',
-      blue: '#58B9F4FF',
-      darkRed: '#C43047FF',
-      glassOverlay: '#2F334899',
-      gradientStop1: '#9469EEFF',
-      gradientStop2: '#C781FCFF',
-      gradient: 'linear-gradient(315deg, #9469EEFF 0%, #C781FCFF 100%)',
-      green: '#6CDF44FF',
-      lightGreen: '#15D864FF',
-      orange: '#EFA947FF',
-      red: '#F9344CFF',
-      trendingBlue: '#2A85E8FF',
-      white: '#2F3348FF',
-      default: '#58B9F4FF',
-      special: '#58B9F4FF'
-    }
-  },
-  matrix: {
-    static: {
-      primary: '#B94CC2FF',
-      secondary: '#7E1BCCFF',
-      white: '#020804FF',
-      staticWhite: '#F6F8FCFF',
-      black: '#000000'
-    },
-    primary: {
-      p100: '#08A008FF',
-      p200: '#09B509FF',
-      p300: '#0BD90BFF',
-      p400: '#0BE40BFF',
-      p500: '#3DF23DFF',
-      default: '#0BD90BFF',
-      primary: '#0BD90BFF'
-    },
-    secondary: {
-      s100: '#00A73CFF',
-      s200: '#00BD44FF',
-      s300: '#00E252FF',
-      s400: '#00ED56FF',
-      s500: '#3CFB6BFF',
-      default: '#00E252FF',
-      secondary: '#00E252FF'
-    },
-    neutral: {
-      n25: '#010B06FF',
-      n50: '#041308FF',
-      n100: '#0A1F0DFF',
-      n150: '#0F2F15FF',
-      n200: '#123F1BFF',
-      n300: '#166527FF',
-      n400: '#1C8736FF',
-      n500: '#1F9E3FFF',
-      n600: '#2EB954FF',
-      n700: '#34D561FF',
-      n800: '#3EF56BFF',
-      n900: '#54FF80FF',
-      n950: '#9BFFA7FF',
-      default: '#3EF56BFF',
-      neutral: '#3EF56BFF'
-    },
-    special: {
-      aiGreen: '#1EEB57FF',
-      background: '#000000FF',
-      blue: '#20E290FF',
-      darkRed: '#D63C5AFF',
-      glassOverlay: '#01010199',
-      gradientStop1: '#6CDF44FF',
-      gradientStop2: '#13C65AFF',
-      gradient: 'linear-gradient(315deg, #6CDF44FF 0%, #13C65AFF 100%)',
-      green: '#2EB954FF',
-      lightGreen: '#0BF90BFF',
-      orange: '#FFA524FF',
-      red: '#F9344CFF',
-      trendingBlue: '#1EEB6FFF',
-      white: '#020804FF',
-      default: '#20E290FF',
-      special: '#20E290FF'
-    }
-  }
-}
-
-export type PrimitiveColorsV2 = typeof primitiveThemeV2.day
-export type PrimaryColorsV2 = keyof PrimitiveColorsV2['primary']
-export type SecondaryColorsV2 = keyof PrimitiveColorsV2['secondary']
-export type NeutralColorsV2 = keyof PrimitiveColorsV2['neutral']
-export type SpecialColorsV2 = keyof PrimitiveColorsV2['special']
-export type StaticColorsV2 = keyof PrimitiveColorsV2['static']
