+++ packages/harmony/src/foundations/color/semantic.ts
@@ -1,15 +1,22 @@
+import { Theme } from '../theme/types'
+
 import type { PrimitiveColors } from './primitive'
 import { primitiveTheme } from './primitive'
 
-const createSemanticTheme = (primitives: PrimitiveColors) => ({
+const createSemanticTheme = (theme: Theme, primitives: PrimitiveColors) => ({
   text: {
-    heading: primitives.special.gradient,
-    default: primitives.neutral.neutral,
+    default: primitives.neutral.n800,
     subdued: primitives.neutral.n400,
     disabled: primitives.neutral.n150,
-    active: primitives.primary.primary,
+    link: primitives.primary.p500,
     accent: primitives.secondary.s300,
-    staticWhite: primitives.static.white,
+    inverse:
+      theme === 'day' ? primitives.neutral.n950 : primitives.special.white,
+
+    // Legacy compatibility
+    heading: primitives.special.gradient,
+    active: primitives.primary.p300,
+    staticWhite: primitives.special.white,
     staticStaticWhite: primitives.static.staticWhite,
     warning: primitives.special.orange,
     danger: primitives.special.red,
@@ -17,13 +24,18 @@ const createSemanticTheme = (primitives: PrimitiveColors) => ({
     special: primitives.special.blue
   },
   icon: {
-    heading: primitives.special.gradient,
-    default: primitives.neutral.neutral,
+    default: primitives.neutral.n800,
     subdued: primitives.neutral.n400,
     disabled: primitives.neutral.n150,
-    active: primitives.primary.primary,
+    link: primitives.primary.p500,
     accent: primitives.secondary.s300,
-    staticWhite: primitives.static.white,
+    inverse:
+      theme === 'day' ? primitives.neutral.n950 : primitives.special.white,
+
+    // Legacy compatibility
+    heading: primitives.special.gradient,
+    active: primitives.primary.p300,
+    staticWhite: primitives.special.white,
     staticStaticWhite: primitives.static.staticWhite,
     warning: primitives.special.orange,
     danger: primitives.special.red,
@@ -31,35 +43,41 @@ const createSemanticTheme = (primitives: PrimitiveColors) => ({
     special: primitives.special.blue
   },
   link: {
-    default: primitives.neutral.neutral,
+    default: primitives.neutral.n800,
     subdued: primitives.neutral.n400,
     visible: primitives.primary.p500,
     inverted: primitives.static.white
   },
   background: {
     default: primitives.special.background,
-    white: primitives.special.white,
     surface1: primitives.neutral.n25,
-    surface2: primitives.neutral.n100,
+    surface2: primitives.neutral.n50,
+    white: primitives.special.white,
+
+    // Legacy compatibility
     accent: primitives.secondary.s300
   },
   border: {
     default: primitives.neutral.n100,
     strong: primitives.neutral.n150,
+    negative: primitives.special.white,
     accent: primitives.secondary.s300
   },
-  focus: { default: primitives.secondary.secondary },
+  focus: {
+    default: primitives.secondary.s300
+  },
   status: {
     error: primitives.special.red,
     warning: primitives.special.orange,
-    success: primitives.special.green
+    success: primitives.special.green,
+    danger: primitives.special.red
   }
 })
 
 export const semanticTheme = {
-  day: createSemanticTheme(primitiveTheme.day),
-  dark: createSemanticTheme(primitiveTheme.dark),
-  matrix: createSemanticTheme(primitiveTheme.matrix)
+  day: createSemanticTheme('day', primitiveTheme.day),
+  dark: createSemanticTheme('dark', primitiveTheme.dark),
+  matrix: createSemanticTheme('matrix', primitiveTheme.matrix)
 }
 
 export type SemanticColors = typeof semanticTheme.day
