+++ packages/harmony/src/foundations/theme/ThemeProvider.tsx
@@ -3,31 +3,24 @@ import { useEffect, type ReactNode } from 'react'
 import { ThemeProvider as EmotionThemeProvider } from '@emotion/react'
 
 import { themes } from './theme'
-import { themesV2 } from './themeV2'
 import type { Theme } from './types'
 
 type ThemeProviderProps = {
   theme: Theme
-  version: 'v1' | 'v2'
   children: ReactNode
 }
 
 export const ThemeProvider = (props: ThemeProviderProps) => {
-  const { children, theme, version } = props
+  const { children, theme } = props
 
   useEffect(() => {
     if (typeof document !== 'undefined') {
-      document.documentElement.setAttribute(
-        'data-theme',
-        version === 'v2' ? `${theme}-v2` : theme
-      )
+      document.documentElement.setAttribute('data-theme', theme)
     }
-  }, [theme, version])
+  }, [theme])
 
   return (
-    <EmotionThemeProvider
-      theme={version === 'v2' ? themesV2[theme] : themes[theme]}
-    >
+    <EmotionThemeProvider theme={themes[theme]}>
       {children}
     </EmotionThemeProvider>
   )
