+++ packages/harmony/src/index.ts
@@ -3,7 +3,6 @@ import './foundations/typography/avenir.css'
 import './foundations/typography/fonts.css'
 import './foundations/spacing/spacing.css'
 import './foundations/color/primitive.css'
-import './foundations/color/primitive-v2.css'
 import './foundations/color/semantic.css'
 import './foundations/motion/motion.css'
 import './foundations/corner-radius/corner-radius.css'
