+++ packages/mobile/src/harmony-native/foundations/theme/ThemeProvider.tsx
@@ -1,12 +1,9 @@
 import type { ReactNode } from 'react'
 
-import { useFeatureFlag } from '@audius/common/hooks'
-import { FeatureFlags } from '@audius/common/services'
 import type { Theme } from '@audius/harmony/src/foundations/theme/types'
 import { ThemeProvider as EmotionThemeProvider } from '@emotion/react'
 
 import { theme } from './theme'
-import { themeV2 } from './themeV2'
 
 type ThemeProviderProps = {
   themeName: Theme
@@ -15,12 +12,9 @@ type ThemeProviderProps = {
 
 export const ThemeProvider = (props: ThemeProviderProps) => {
   const { children, themeName } = props
-  const { isEnabled: isThemeV2Enabled } = useFeatureFlag(FeatureFlags.THEME_V2)
 
   return (
-    <EmotionThemeProvider
-      theme={isThemeV2Enabled ? themeV2[themeName] : theme[themeName]}
-    >
+    <EmotionThemeProvider theme={theme[themeName]}>
       {children}
     </EmotionThemeProvider>
   )
