+++ packages/mobile/src/styles/makeAnimations.ts
@@ -1,6 +1,4 @@
-import { useFeatureFlag } from '@audius/common/hooks'
 import { Theme } from '@audius/common/models'
-import { FeatureFlags } from '@audius/common/services'
 
 import type { ThemeColors } from 'app/utils/theme'
 import {
@@ -9,7 +7,6 @@ import {
   matrixTheme,
   defaultTheme
 } from 'app/utils/theme'
-import { themeColorsV2 } from 'app/utils/themeV2'
 
 type AnimationCreatorConfig = { palette: ThemeColors; type: Theme }
 
@@ -31,29 +28,8 @@ export const makeAnimations = <TReturn>(
     })
   }
 
-  const themedAnimationsV2 = {
-    [Theme.DEFAULT]: animationCreator({
-      palette: themeColorsV2[Theme.DEFAULT],
-      type: Theme.DEFAULT
-    }),
-    [Theme.DARK]: animationCreator({
-      palette: themeColorsV2[Theme.DARK],
-      type: Theme.DARK
-    }),
-    [Theme.MATRIX]: animationCreator({
-      palette: themeColorsV2[Theme.MATRIX],
-      type: Theme.MATRIX
-    })
-  }
-
   return function useAnimations(): TReturn {
     const themeVariant = useThemeVariant()
-    const { isEnabled: isThemeV2Enabled } = useFeatureFlag(
-      FeatureFlags.THEME_V2
-    )
-
-    return isThemeV2Enabled
-      ? themedAnimationsV2[themeVariant]
-      : themedAnimations[themeVariant]
+    return themedAnimations[themeVariant]
   }
 }
