+++ packages/mobile/src/styles/makeStyles.ts
@@ -1,6 +1,4 @@
-import { useFeatureFlag } from '@audius/common/hooks'
 import { Theme } from '@audius/common/models'
-import { FeatureFlags } from '@audius/common/services'
 import type { ImageStyle, TextStyle, ViewStyle } from 'react-native'
 import { StyleSheet } from 'react-native'
 
@@ -11,7 +9,6 @@ import {
   darkTheme,
   useThemeVariant
 } from 'app/utils/theme'
-import { themeColorsV2 } from 'app/utils/themeV2'
 
 import { spacing } from './spacing'
 import { typography } from './typography'
@@ -58,50 +55,14 @@ export const makeStyles = <T extends Record<string, StyleTypes>>(
     })
   )
 
-  const defaultStylesheetV2 = StyleSheet.create(
-    styles({
-      type: Theme.DEFAULT,
-      palette: themeColorsV2[Theme.DEFAULT],
-      ...baseOptions
-    })
-  )
-
-  const darkStylesheetV2 = StyleSheet.create(
-    styles({
-      type: Theme.DARK,
-      palette: themeColorsV2[Theme.DARK],
-      ...baseOptions
-    })
-  )
-
-  const matrixStylesheetV2 = StyleSheet.create(
-    styles({
-      type: Theme.MATRIX,
-      palette: themeColorsV2[Theme.MATRIX],
-      ...baseOptions
-    })
-  )
-
   const themedStylesheets = {
     [Theme.DEFAULT]: defaultStylesheet,
     [Theme.DARK]: darkStylesheet,
     [Theme.MATRIX]: matrixStylesheet
   }
 
-  const themedStylesheetsV2 = {
-    [Theme.DEFAULT]: defaultStylesheetV2,
-    [Theme.DARK]: darkStylesheetV2,
-    [Theme.MATRIX]: matrixStylesheetV2
-  }
-
   return function useStyles() {
     const themeVariant = useThemeVariant()
-    const { isEnabled: isThemeV2Enabled } = useFeatureFlag(
-      FeatureFlags.THEME_V2
-    )
-
-    return isThemeV2Enabled
-      ? themedStylesheetsV2[themeVariant]
-      : themedStylesheets[themeVariant]
+    return themedStylesheets[themeVariant]
   }
 }
