+++ packages/web/src/app/ThemeProvider.tsx
@@ -1,8 +1,6 @@
 import { ReactNode, useEffect } from 'react'
 
-import { useFeatureFlag } from '@audius/common/hooks'
 import { Theme, SystemAppearance } from '@audius/common/models'
-import { FeatureFlags } from '@audius/common/services'
 import { themeActions, themeSelectors } from '@audius/common/store'
 import { ThemeProvider as HarmonyThemeProvider } from '@audius/harmony'
 import { useDispatch } from 'react-redux'
@@ -48,7 +46,6 @@ export const ThemeProvider = (props: ThemeProviderProps) => {
   const { children } = props
   const harmonyTheme = useSelector(selectHarmonyTheme)
   const dispatch = useDispatch()
-  const { isEnabled: isThemeV2Enabled } = useFeatureFlag(FeatureFlags.THEME_V2)
 
   useEffect(() => {
     if (typeof window === 'undefined' || !window.matchMedia) {
@@ -75,11 +72,6 @@ export const ThemeProvider = (props: ThemeProviderProps) => {
   }, [dispatch])
 
   return (
-    <HarmonyThemeProvider
-      theme={harmonyTheme}
-      version={isThemeV2Enabled ? 'v2' : 'v1'}
-    >
-      {children}
-    </HarmonyThemeProvider>
+    <HarmonyThemeProvider theme={harmonyTheme}>{children}</HarmonyThemeProvider>
   )
 }
